fix: created custom exception and replaced floats with doubles after cr

+++ src/main/java/com/david/nonogramgeneratorapi/dtos/nonogramGenerationRequestDto.java
@@ -13,5 +13,5 @@ import java.io.Serializable;
 public class nonogramGenerationRequestDto implements Serializable {
     private String imageBase64;
     private Difficulty difficulty;
-    private float pixelHighlightValue;
+    private double pixelHighlightValue;
 }
